+++ axelor-production/src/main/java/com/axelor/apps/production/web/ProdProcessController.java
@@ -17,23 +17,30 @@
  */
 package com.axelor.apps.production.web;
 
+import java.util.List;
 import com.axelor.apps.ReportFactory;
 import com.axelor.apps.production.db.BillOfMaterial;
 import com.axelor.apps.production.db.ProdProcess;
 import com.axelor.apps.production.db.repo.BillOfMaterialRepository;
+import com.axelor.apps.production.db.repo.ProdProcessRepository;
 import com.axelor.apps.production.report.IReport;
 import com.axelor.apps.production.service.ProdProcessService;
 import com.axelor.exception.AxelorException;
+import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.axelor.meta.schema.actions.ActionView;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
+import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 
 public class ProdProcessController {
 	
 	@Inject
 	protected ProdProcessService prodProcessService;
+	
+	@Inject
+	ProdProcessRepository prodProcessRepo;
 	
 	public void validateProdProcess(ActionRequest request, ActionResponse response) throws AxelorException{
 		ProdProcess prodProcess = request.getContext().asType(ProdProcess.class);
@@ -75,5 +82,41 @@ public class ProdProcessController {
 				.define(prodProcessLabel)
 				.add("html", fileLink).map());
 		
+	}
+	
+	public void checkOriginalProductionProcess(ActionRequest request, ActionResponse response) {
+
+		ProdProcess prodProcess = prodProcessRepo.find(request.getContext().asType(ProdProcess.class).getId());
+
+		List<ProdProcess> prodProcessSet = Lists.newArrayList();
+		prodProcessSet = prodProcessRepo.all().filter("self.originalProdProcess = :origin").bind("origin", prodProcess)
+				.fetch();
+		String message;
+
+		if (!prodProcessSet.isEmpty()) {
+
+			String existingVersions = "";
+			for (ProdProcess prodProcessVersion : prodProcessSet) {
+				existingVersions += "<li>" + prodProcessVersion.getFullName() + "</li>";
+			}
+			message = String.format(I18n.get(
+					"This production process already has the following versions : <br/><ul> %s </ul>And these versions may also have ones. Do you still wish to create a new one ?"),
+					existingVersions);
+		} else {
+			message = I18n.get("Do you really wish to create a new version of this production process ?");
+		}
+
+		response.setAlert(message);
+	}
+
+	public void generateNewVersion(ActionRequest request, ActionResponse response) {
+
+		ProdProcess prodProcess = prodProcessRepo.find(request.getContext().asType(ProdProcess.class).getId());
+
+		ProdProcess copy = prodProcessService.generateNewVersion(prodProcess);
+
+		response.setView(ActionView.define("Production process").model(ProdProcess.class.getName())
+				.add("form", "prod-process-form").add("grid", "prod-process-grid")
+				.context("_showRecord", String.valueOf(copy.getId())).map());
 	}
 }